fix: ASN recategorization and database optimizations
- Fix ASN startup logic to recategorize when >50% are unknown
- Add detailed logging for ipverse integration and recategorization
- Optimize database: convert IPs to INET, ports to SMALLINT
- Add optimized indexes: trigram for player search, GIN for tags, partial for flags
- Update entity models and repositories for type safety
- Document ASN system and force-asn-import usage in README

+++ src/models/entities/server_history.rs
@@ -5,8 +5,8 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub ip: String,
-    #[sea_orm(primary_key)]
-    pub port: i32,
+    #[sea_orm(primary_key, column_type = "SmallInteger")]
+    pub port: i16,
     #[sea_orm(primary_key)]
     pub timestamp: DateTime,
     pub players_online: i32,
